+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ElementWrapper.findall` and `ElementWrapper.find` methods
 - Additional CPHD consistency checks for Support Arrays and associated unit tests
 - Methods for converting to & from image area coordinates in `sarkit.cphd`
+- Methods for computing "reference times" in `sarkit.cphd`
 
 
 ## [1.5.0] - 2026-03-04