+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.6.0] - 2026-03-30
+
 ### Added
 - `ElementWrapper.findall` and `ElementWrapper.find` methods
 - Additional CPHD consistency checks for Support Arrays and associated unit tests
@@ -230,7 +233,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Limited SICD Reading and Writing
 - Limited SIDD NITF Reading and Writing
 
-[unreleased]: https://github.com/ValkyrieSystems/sarkit/compare/v1.5.0...HEAD
+[unreleased]: https://github.com/ValkyrieSystems/sarkit/compare/v1.6.0...HEAD
+[1.6.0]: https://github.com/ValkyrieSystems/sarkit/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/ValkyrieSystems/sarkit/compare/v1.4.1...v1.5.0
 [1.4.1]: https://github.com/ValkyrieSystems/sarkit/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/ValkyrieSystems/sarkit/compare/v1.3.0...v1.4.0